remove constant pool type, no longer needed
